requirements.txt
aiofiles==23.2.1
aiogram==3.12.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
APScheduler==3.10.4
attrs==24.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.3.2
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
idna==3.8
magic-filter==1.0.12
multidict==6.0.5
outcome==1.3.0.post0
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv
PySocks==1.7.1
pytz==2024.1
requests==2.32.3
selenium==4.24.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.34
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.2
websocket-client==1.8.0
websockets==8.1
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.8
